+++ src/test/kotlin/com/dooingle/domain/dooingle/service/DooingleServiceUnitTest.kt
@@ -38,14 +38,16 @@ class DooingleServiceUnitTest : AnnotationSpec() {
     private val mockDooingleRepository = mockk<DooingleRepository>()
     private val mockSocialUserRepository = mockk<SocialUserRepository>()
     private val mockCatchRepository = mockk<CatchRepository>()
+    private val mockDooingleCountRepository = mockk<DooingleCountRepository>()
+    private val mockNotificationService = mockk<NotificationService>()
     private val mockDistributedLock = mockk<DistributedLock>()
-    private val mockDooingleAddService = mockk<DooingleAddService>()
     private val dooingleService = DooingleService(
         dooingleRepository = mockDooingleRepository,
         socialUserRepository = mockSocialUserRepository,
         catchRepository = mockCatchRepository,
-        distributedLock = mockDistributedLock,
-        dooingleAddService = mockDooingleAddService,
+        dooingleCountRepository = mockDooingleCountRepository,
+        notificationService = mockNotificationService,
+        distributedLock = mockDistributedLock
     )
 
     lateinit var owner: SocialUser
